# JSON API: avoid excessive iterations in benchmarks