Also:
* fix db connection setup - the combination of annotations and inheritance meant it was trying to setup the trial twice, causing Postgres benchmarks to fail
* log when an error occurs during test setup
* add logback resources to the benchmarks, to enable configuration of logging, and avoid dumping all the debug logs by default

With these changes, on my 20 core 64G linux laptop, the benchmarks now all run in under 6 mins

```
# Run complete. Total time: 00:05:51

Benchmark                          (batchSize)  (batches)  (extraParties)  (extraPayloadValues)  (extraTemplates)  (numContracts)  Mode  Cnt  Score    Error  Units
InsertBenchmarkOracle.run                 1000          1             N/A                   N/A               N/A            1000  avgt       0.127            s/op
InsertBenchmarkOracle.run                 1000        100             N/A                   N/A               N/A            1000  avgt       0.134            s/op
InsertBenchmarkPostgres.run               1000          1             N/A                   N/A               N/A            1000  avgt       0.040            s/op
InsertBenchmarkPostgres.run               1000        100             N/A                   N/A               N/A            1000  avgt       0.035            s/op
QueryBenchmarkOracle.run                  1000        N/A               1                   N/A                 1             N/A  avgt   10  0.008 ±  0.001   s/op
QueryBenchmarkOracle.run                  1000        N/A               1                   N/A                10             N/A  avgt   10  0.009 ±  0.001   s/op
QueryBenchmarkOracle.run                  1000        N/A              10                   N/A                 1             N/A  avgt   10  0.015 ±  0.001   s/op
QueryBenchmarkOracle.run                  1000        N/A              10                   N/A                10             N/A  avgt   10  0.008 ±  0.001   s/op
QueryBenchmarkPostgres.run                1000        N/A               1                   N/A                 1             N/A  avgt   10  0.002 ±  0.001   s/op
QueryBenchmarkPostgres.run                1000        N/A               1                   N/A                10             N/A  avgt   10  0.002 ±  0.001   s/op
QueryBenchmarkPostgres.run                1000        N/A              10                   N/A                 1             N/A  avgt   10  0.003 ±  0.001   s/op
QueryBenchmarkPostgres.run                1000        N/A              10                   N/A                10             N/A  avgt   10  0.003 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run           1000        N/A               1                     1               N/A             N/A  avgt   10  0.011 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run           1000        N/A               1                    10               N/A             N/A  avgt   10  0.047 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run           1000        N/A              10                     1               N/A             N/A  avgt   10  0.011 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run           1000        N/A              10                    10               N/A             N/A  avgt   10  0.051 ±  0.005   s/op
QueryPayloadBenchmarkPostgres.run         1000        N/A               1                     1               N/A             N/A  avgt   10  0.006 ±  0.003   s/op
QueryPayloadBenchmarkPostgres.run         1000        N/A               1                    10               N/A             N/A  avgt   10  0.011 ±  0.001   s/op
QueryPayloadBenchmarkPostgres.run         1000        N/A              10                     1               N/A             N/A  avgt   10  0.007 ±  0.005   s/op
QueryPayloadBenchmarkPostgres.run         1000        N/A              10                    10               N/A             N/A  avgt   10  0.032 ±  0.001   s/op

```

without this change, it took over 9 hours in total and none of the Postgres benchmarks were successful.

```
# Run complete. Total time: 09:13:36

Benchmark                        (batchSize)  (batches)  (extraParties)  (extraPayloadValues)  (extraTemplates)  (numContracts)  Mode  Cnt  Score    Error  Units
InsertBenchmarkOracle.run               1000          1             N/A                   N/A               N/A            1000  avgt  200  0.109 ±  0.002   s/op
InsertBenchmarkOracle.run               1000          3             N/A                   N/A               N/A            1000  avgt  200  0.114 ±  0.005   s/op
InsertBenchmarkOracle.run               1000          5             N/A                   N/A               N/A            1000  avgt  200  0.120 ±  0.018   s/op
InsertBenchmarkOracle.run               1000          7             N/A                   N/A               N/A            1000  avgt  200  0.118 ±  0.012   s/op
InsertBenchmarkOracle.run               1000          9             N/A                   N/A               N/A            1000  avgt  200  0.126 ±  0.023   s/op
QueryBenchmarkOracle.run                1000        N/A               1                   N/A                 1             N/A  avgt  200  0.008 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               1                   N/A                 5             N/A  avgt  200  0.009 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               1                   N/A                 9             N/A  avgt  200  0.009 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               5                   N/A                 1             N/A  avgt  200  0.011 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               5                   N/A                 5             N/A  avgt  200  0.013 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               5                   N/A                 9             N/A  avgt  200  0.013 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               9                   N/A                 1             N/A  avgt  200  0.014 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               9                   N/A                 5             N/A  avgt  200  0.016 ±  0.001   s/op
QueryBenchmarkOracle.run                1000        N/A               9                   N/A                 9             N/A  avgt  200  0.017 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A               1                     1               N/A             N/A  avgt  200  0.011 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A               1                   100               N/A             N/A  avgt  200  0.387 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A               1                   100               N/A             N/A  avgt  200  0.387 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A              10                     1               N/A             N/A  avgt  200  0.011 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A              10                   100               N/A             N/A  avgt  200  0.417 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A              10                   100               N/A             N/A  avgt  200  0.417 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A             100                     1               N/A             N/A  avgt  200  0.011 ±  0.001   s/op
QueryPayloadBenchmarkOracle.run         1000        N/A             100                   100               N/A             N/A  avgt   40  0.475 ±  0.001   s/op
```